refactor: 클라이언트 분리
