+++ src/worker/application/client.ts
@@ -1,0 +1,5 @@
+export interface IWorkerClient {
+    readByOptions(query: any): Promise<any>;
+
+    updatePartial(query: any): Promise<any>;
+}
