+++ src/worker/application/dispatch.ts
@@ -1,4 +1,4 @@
-import NotificationEntity from "@notification/domain/entity";
+import { NotificationEntity } from "./dto";
 
 export default async function dispatch(notifications: NotificationEntity[]) {
     console.log(notifications);
