--- src/worker/dto.ts
@@ -1,5 +1,0 @@
-export enum NotificationStatus {
-    Pending = "Pending",
-    Sent = "Sent",
-    Failed = "Failed",
-}
