+++ src/worker/main.ts
@@ -1,7 +1,7 @@
 import { NestFactory } from "@nestjs/core";
 
 import { WorkerModule } from "./module";
-import { WorkerService } from "./service";
+import { WorkerService } from "./application/service";
 
 const WORKER_PORT = process.env.WORKER_PORT || 3001;
